Feat/topup

+++ src/main/java/id/ac/ui/cs/advprog/bepayment/controller/TopUpController.java
@@ -1,13 +1,22 @@
 package id.ac.ui.cs.advprog.bepayment.controller;
 
+import id.ac.ui.cs.advprog.bepayment.config.JwtAuthFilter;
 import id.ac.ui.cs.advprog.bepayment.model.TopUp;
 import id.ac.ui.cs.advprog.bepayment.pojos.TopUpRequest;
 import id.ac.ui.cs.advprog.bepayment.service.TopUpService;
+import io.jsonwebtoken.ExpiredJwtException;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 
+import java.io.IOException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -22,9 +31,30 @@ public class TopUpController {
     @Autowired
     private TopUpService topUpService;
 
-    @PostMapping("/create")
-    public CompletableFuture<ResponseEntity<Map<String, Object>>> createTopUp(@RequestBody TopUpRequest topUpRequest) {
+    @Autowired
+    private JwtAuthFilter jwtAuthFilter;
+
+    @RequestMapping(value = "/create", method = RequestMethod.POST)
+    public CompletableFuture<ResponseEntity<Map<String, Object>>> createTopUp(@RequestHeader(value = "Authorization") String token, @RequestBody TopUpRequest topUpRequest) {
         Map<String, Object> response = new HashMap<>();
+        String role = null;
+        try {
+            role = jwtAuthFilter.filterToken(token);
+        }catch (ExpiredJwtException e){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "JWT token has expired");
+        }catch (Exception e){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "Invalid JWT token");
+
+        }
+
+        if (role == null){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "Login First");
+            return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.FORBIDDEN).body(response));
+        }
+
         CompletableFuture<TopUp> createdTopUpFuture = topUpService.createTopUp(topUpRequest);
 
         return createdTopUpFuture.thenApply(createdTopUp -> {
@@ -41,8 +71,27 @@ public class TopUpController {
 
 
     @DeleteMapping("/")
-    public CompletableFuture<ResponseEntity<Map<String, Object>>> deleteAllTopUp() {
+    public CompletableFuture<ResponseEntity<Map<String, Object>>> deleteAllTopUp(@RequestHeader(value = "Authorization") String token) {
         Map<String, Object> response = new HashMap<>();
+
+        String role = null;
+        try {
+            role = jwtAuthFilter.filterToken(token);
+        }catch (ExpiredJwtException e){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "JWT token has expired");
+        }catch (Exception e){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "Invalid JWT token");
+
+        }
+
+        if (role == null){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "Login First");
+            return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.FORBIDDEN).body(response));
+        }
+
         CompletableFuture<Object> deletionFuture = topUpService.deleteAllTopUp();
 
         return deletionFuture.thenApply(deleted -> {
@@ -59,8 +108,27 @@ public class TopUpController {
 
 
     @DeleteMapping("/{topUpId}/delete")
-    public CompletableFuture<ResponseEntity<Map<String, Object>>> deleteTopUpById(@PathVariable("topUpId") String topUpId) {
+    public CompletableFuture<ResponseEntity<Map<String, Object>>> deleteTopUpById(@RequestHeader(value = "Authorization") String token, @PathVariable("topUpId") String topUpId) {
         Map<String, Object> response = new HashMap<>();
+
+        String role = null;
+        try {
+            role = jwtAuthFilter.filterToken(token);
+        }catch (ExpiredJwtException e){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "JWT token has expired");
+        }catch (Exception e){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "Invalid JWT token");
+
+        }
+
+        if (role == null){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "Login First");
+            return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.FORBIDDEN).body(response));
+        }
+
         return topUpService.deleteTopUpById(topUpId)
                 .thenApply(deleted -> {
                     if (deleted) {
@@ -83,8 +151,26 @@ public class TopUpController {
 
 
     @PutMapping("/{topUpId}/cancel")
-    public CompletableFuture<ResponseEntity<Map<String, Object>>> cancelTopUp(@PathVariable("topUpId") String topUpId) {
+    public CompletableFuture<ResponseEntity<Map<String, Object>>> cancelTopUp(@RequestHeader(value = "Authorization") String token, @PathVariable("topUpId") String topUpId) {
         Map<String, Object> response = new HashMap<>();
+
+        String role = null;
+        try {
+            role = jwtAuthFilter.filterToken(token);
+        }catch (ExpiredJwtException e){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "JWT token has expired");
+        }catch (Exception e){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "Invalid JWT token");
+
+        }
+
+        if (role == null){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "Login First");
+            return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.FORBIDDEN).body(response));
+        }
 
         return topUpService.cancelTopUp(topUpId)
                 .thenApply(cancelled -> {
@@ -108,8 +194,30 @@ public class TopUpController {
 
 
     @PutMapping("/{topUpId}/confirm")
-    public CompletableFuture<ResponseEntity<Map<String, Object>>> confirmTopUp(@PathVariable("topUpId") String topUpId) {
+    public CompletableFuture<ResponseEntity<Map<String, Object>>> confirmTopUp(@RequestHeader(value = "Authorization") String token, @PathVariable("topUpId") String topUpId) {
         Map<String, Object> response = new HashMap<>();
+
+        String role = null;
+        try {
+            role = jwtAuthFilter.filterToken(token);
+        }catch (ExpiredJwtException e){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "JWT token has expired");
+        }catch (Exception e){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "Invalid JWT token");
+
+        }
+
+        if (role == null){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "Login First");
+            return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.FORBIDDEN).body(response));
+        }else if (role.equals("REGULAR")){
+            response.put("code", HttpStatus.FORBIDDEN.value());
+            response.put("message", "You cant do this");
+            return CompletableFuture.completedFuture(ResponseEntity.status(HttpStatus.FORBIDDEN).body(response));
+        }
 
         return topUpService.confirmTopUp(topUpId)
                 .thenApply(confirmed -> {
@@ -134,18 +242,17 @@ public class TopUpController {
 
     @GetMapping("/")
     public CompletableFuture<ResponseEntity<List<TopUp>>> getAllTopUps() {
+        Map<String, Object> response = new HashMap<>();
+
         CompletableFuture<List<TopUp>> topUpsFuture = topUpService.findAll();
         return topUpsFuture.thenApplyAsync(topUps -> ResponseEntity.ok(topUps))
                 .exceptionally(exception -> {
-                    Map<String, Object> response = new HashMap<>();
                     response.put("code", HttpStatus.INTERNAL_SERVER_ERROR.value());
                     response.put("error", exception.getCause() != null ? exception.getCause().getMessage() : "Unknown error");
                     response.put("message", "Something went wrong with the server");
                     return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Collections.emptyList());
                 });
     }
-
-
 
 
     @GetMapping("/waiting")
@@ -174,13 +281,11 @@ public class TopUpController {
             }
             return ResponseEntity.ok(topUp);
         }).exceptionally(exception -> {
-            // Create a map containing the error details
             Map<String, Object> errorResponse = new HashMap<>();
             errorResponse.put("code", HttpStatus.INTERNAL_SERVER_ERROR.value());
             errorResponse.put("error", "Internal Server Error");
             errorResponse.put("message", "Something went wrong with the server");
 
-            // Return the error response as the body of the response entity
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
         });
     }
@@ -193,18 +298,13 @@ public class TopUpController {
             successResponse.put("topUps", topUps);
             return ResponseEntity.ok(successResponse);
         }).exceptionally(exception -> {
-            // Create a map containing the error details
             Map<String, Object> errorResponse = new HashMap<>();
             errorResponse.put("code", HttpStatus.INTERNAL_SERVER_ERROR.value());
             errorResponse.put("error", "Internal Server Error");
             errorResponse.put("message", "Something went wrong with the server");
 
-            // Return the error response as the body of the response entity
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
         });
     }
-
-
-
 
 }
